requirements.txt
langchain
langchain_core
langchain_community
langchain-cohere

streamlit
python-dotenv